Extract BSD OperatingSystem common bases (JNA/FFM)
The FreeBSD, OpenBSD, and DragonFly OperatingSystem classes each duplicated
their entire body between the JNA and FFM backends (the method signatures were
identical; only the native calls and backend factory types differed).

Following the Solaris #3367 template, extract an abstract per-platform base in
oshi-common holding the ps-based process enumeration, process/thread counts,
services, and system uptime/boot time. The JNA and FFM subclasses are reduced to
the genuinely backend-specific bits: the OSProcess factory (a new createProcess
hook), the FileSystem/NetworkParams/InternetProtocolStats factories, getpid/
thread-id, the sysctl version query, and who sessions.

Per-platform notes:
- FreeBSD/DragonFly read boot time from a native sysctl struct, so the base
  declares an abstract queryBootTime() hook (memoized); OpenBSD text-parses it,
  so its base implements queryBootTime() concretely.
- OpenBSD's InternetProtocolStats and NetworkParams are common (non-native)
  classes, so they move into the base; FreeBSD/DragonFly keep their JNA/FFM
  factory overrides. DragonFly reuses FreeBSD's filesystem/IP/network impls and
  keeps its kernel-thread (pid <= 0) filter.

NetBSD already has a single common OperatingSystem class and is unchanged. The
JNA/FFM subclasses shrink from ~225-238 lines each to ~55-90.
