+++ oshi-common/src/main/java/oshi/software/common/os/unix/openbsd/OpenBsdOperatingSystem.java
@@ -1,0 +1,209 @@
+/*
+ * Copyright 2021-2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+package oshi.software.common.os.unix.openbsd;
+
+import static oshi.software.os.OSService.State.RUNNING;
+import static oshi.software.os.OSService.State.STOPPED;
+import static oshi.util.Memoizer.memoize;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import oshi.annotation.concurrent.ThreadSafe;
+import oshi.software.common.AbstractOperatingSystem;
+import oshi.software.common.os.unix.bsd.BsdPsKeyword;
+import oshi.software.os.InternetProtocolStats;
+import oshi.software.os.NetworkParams;
+import oshi.software.os.OSProcess;
+import oshi.software.os.OSService;
+import oshi.software.os.OSThread;
+import oshi.util.ExecutingCommand;
+import oshi.util.ParseUtil;
+
+/**
+ * Abstract base shared by the OpenBSD OperatingSystem implementations (JNA and FFM). Holds the {@code ps}-based process
+ * enumeration, process/thread counts, services, internet protocol stats, network params, and system uptime/boot time.
+ * The native bits (the {@link OSProcess} and {@code FileSystem} factories, {@code getpid}/{@code getthrid}, and the
+ * sysctl version query) are provided by the JNA and FFM subclasses.
+ */
+@ThreadSafe
+public abstract class OpenBsdOperatingSystem extends AbstractOperatingSystem {
+
+    private static final Logger LOG = LoggerFactory.getLogger(OpenBsdOperatingSystem.class);
+
+    private final Supplier<Long> bootTime = memoize(this::queryBootTime);
+
+    @Override
+    public String queryManufacturer() {
+        return "Unix/BSD";
+    }
+
+    @Override
+    protected int queryBitness(int jvmBitness) {
+        if (jvmBitness < 64 && ExecutingCommand.getFirstAnswer("uname -m").indexOf("64") == -1) {
+            return jvmBitness;
+        }
+        return 64;
+    }
+
+    @Override
+    public InternetProtocolStats getInternetProtocolStats() {
+        return new OpenBsdInternetProtocolStats();
+    }
+
+    @Override
+    public NetworkParams getNetworkParams() {
+        return new OpenBsdNetworkParams();
+    }
+
+    @Override
+    public List<OSProcess> queryAllProcesses() {
+        return getProcessListFromPS(-1);
+    }
+
+    @Override
+    public List<OSProcess> queryChildProcesses(int parentPid) {
+        List<OSProcess> allProcs = queryAllProcesses();
+        Set<Integer> descendantPids = getChildrenOrDescendants(allProcs, parentPid, false);
+        return allProcs.stream().filter(p -> descendantPids.contains(p.getProcessID())).collect(Collectors.toList());
+    }
+
+    @Override
+    public List<OSProcess> queryDescendantProcesses(int parentPid) {
+        List<OSProcess> allProcs = queryAllProcesses();
+        Set<Integer> descendantPids = getChildrenOrDescendants(allProcs, parentPid, true);
+        return allProcs.stream().filter(p -> descendantPids.contains(p.getProcessID())).collect(Collectors.toList());
+    }
+
+    @Override
+    public OSProcess getProcess(int pid) {
+        List<OSProcess> procs = getProcessListFromPS(pid);
+        if (procs.isEmpty()) {
+            return null;
+        }
+        return procs.get(0);
+    }
+
+    private List<OSProcess> getProcessListFromPS(int pid) {
+        List<OSProcess> procs = new ArrayList<>();
+        // https://man.openbsd.org/ps#KEYWORDS
+        // missing are threadCount and kernelTime which is included in cputime
+        String psCommand = "ps -awwxo " + OpenBsdOSProcess.PS_COMMAND_ARGS;
+        if (pid >= 0) {
+            psCommand += " -p " + pid;
+        }
+        List<String> procList = ExecutingCommand.runNative(psCommand);
+        if (procList.isEmpty() || procList.size() < 2) {
+            return procs;
+        }
+
+        // remove header row
+        procList.remove(0);
+        // Fill list
+        for (String proc : procList) {
+            Map<BsdPsKeyword, String> psMap = ParseUtil.stringToEnumMap(BsdPsKeyword.class,
+                    OpenBsdOSProcess.PS_KEYWORDS, proc.trim(), ' ');
+            // Check if last (thus all) value populated
+            if (psMap.containsKey(BsdPsKeyword.ARGS)) {
+                procs.add(createProcess(pid < 0 ? ParseUtil.parseIntOrDefault(psMap.get(BsdPsKeyword.PID), 0) : pid,
+                        psMap));
+            }
+        }
+        return procs;
+    }
+
+    @Override
+    public int getProcessCount() {
+        List<String> procList = ExecutingCommand.runNative("ps -axo pid");
+        if (!procList.isEmpty()) {
+            // Subtract 1 for header
+            return procList.size() - 1;
+        }
+        return 0;
+    }
+
+    @Override
+    public OSThread getCurrentThread() {
+        OSProcess proc = getCurrentProcess();
+        final int tid = getThreadId();
+        return proc.getThreadDetails().stream().filter(t -> t.getThreadId() == tid).findFirst()
+                .orElse(new OpenBsdOSThread(proc.getProcessID(), tid));
+    }
+
+    @Override
+    public int getThreadCount() {
+        // -H "Also display information about kernel visible threads"
+        // -k "Also display information about kernel threads"
+        // column TID holds thread ID
+        List<String> threadList = ExecutingCommand.runNative("ps -axHo tid");
+        if (!threadList.isEmpty()) {
+            // Subtract 1 for header
+            return threadList.size() - 1;
+        }
+        return 0;
+    }
+
+    @Override
+    public long getSystemUptime() {
+        return System.currentTimeMillis() / 1000 - getSystemBootTime();
+    }
+
+    @Override
+    public long getSystemBootTime() {
+        return bootTime.get();
+    }
+
+    private long queryBootTime() {
+        // Boot time will be the first consecutive string of digits.
+        return ParseUtil.parseLongOrDefault(
+                ExecutingCommand.getFirstAnswer("sysctl -n kern.boottime").split(",")[0].replaceAll("\\D", ""),
+                System.currentTimeMillis() / 1000);
+    }
+
+    @Override
+    public List<OSService> getServices() {
+        // Get running services
+        List<OSService> services = new ArrayList<>();
+        Set<String> running = new HashSet<>();
+        for (OSProcess p : getChildProcesses(1, ProcessFiltering.ALL_PROCESSES, ProcessSorting.PID_ASC, 0)) {
+            OSService s = new OSService(p.getName(), p.getProcessID(), RUNNING);
+            services.add(s);
+            running.add(p.getName());
+        }
+        // Get Directories for stopped services
+        File dir = new File("/etc/rc.d");
+        File[] listFiles;
+        if (dir.exists() && dir.isDirectory() && (listFiles = dir.listFiles()) != null) {
+            for (File f : listFiles) {
+                String name = f.getName();
+                if (!running.contains(name)) {
+                    OSService s = new OSService(name, 0, STOPPED);
+                    services.add(s);
+                }
+            }
+        } else {
+            LOG.error("Directory: /etc/rc.d does not exist");
+        }
+        return services;
+    }
+
+    /**
+     * Creates a backend-specific {@link OSProcess} from a parsed {@code ps} row.
+     *
+     * @param pid   the process ID
+     * @param psMap the parsed {@code ps} columns for this process
+     * @return a new OSProcess
+     */
+    protected abstract OSProcess createProcess(int pid, Map<BsdPsKeyword, String> psMap);
+}
